Added in pause function, beautified
Ran through a formatter as some things got too messy. Added in a 'pause'
functionallity, in which occurs when all three buttons are pressed. This
disables auto refresh and any other button checks. This allows us to
keep our slightly intensive LCD program formatting, letting us pause
when we're starting a match to get rid of load on cpu.