Updated drive, auton, claw, lcd pages, and ports
Changed driveWithLogicHolonomic to driveHolonomicWithLogic to better fit
it's purpose (calls driveHolonomic after doing some logic on inputs).
Removed unused "reverse" boolean from driveHolonomicWithLogic. Added
getter for our gyro so it can be accessed outside of init. Added sensor
disables in auton if we're in a game. Renamed RPMToMotor and motorToRPM
in claw.c to have "claw" on the end to avoid conflicts with sensors.c.
Added LCD pages to monitor sensors. Added in claw ports and line sensor
ports.