Claw Code
This is Mihchael typing, and I think I am going to regret taking
responsiblility for this commit.  I am really not all that confident
that the claw code will work.  Still if we look in blame to find who
sent the code that did not work I feel morally responsible to take the
blame